% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themebg.R
\name{theme_bg}
\alias{theme_bg}
\title{Custom ggplot2 theme}
\usage{
theme_bg(base_size = 11, base_family = "")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}
}
\description{
This is a custom ggplot2 theme loosely based on recommendations of Edward
Tufte, Stephen Few, and others. By default, the panel background is blank,
gridlines are removed, and axis colors are grey to minimize distraction from
the data.
}
