% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themebg.R
\name{theme_bg_ppt}
\alias{theme_bg_ppt}
\title{Custom ggplot2 theme for PowerPoint}
\usage{
theme_bg_ppt(base_size = 11, base_family = "", xticks = TRUE,
  yticks = TRUE, family = "Calibri")
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{xticks}{logical indicating whether tick marks should be displayed on
the x-axis}

\item{yticks}{logical indicating whether tick marks should be displayed on
the y-axis}

\item{family}{string with a font family name, defaults to Calibri}
}
\description{
This is a custom ggplot2 theme loosely based on recommendations of Edward
Tufte, Stephen Few, and others. By default, the panel background is blank,
gridlines are removed, and axis colors are grey to minimize distraction from
the data. This theme is designed for use in PowerPoint with a white
background.
}
