% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themebg.R
\name{theme_bg_print}
\alias{theme_bg_print}
\title{Custom ggplot2 theme for black and white printing}
\usage{
theme_bg_print(base_size = 11, base_family = "", xticks = TRUE,
  yticks = TRUE)
}
\arguments{
\item{base_size}{base font size}

\item{base_family}{base font family}

\item{xticks}{logical indicating whether tick marks should be displayed on
the x-axis}

\item{yticks}{logical indicating whether tick marks should be displayed on
the y-axis}
}
\description{
This is a custom ggplot2 theme loosely based on recommendations of Edward
Tufte, Stephen Few, and others. The panel background is blank and gridlines
are removed. In contrast to theme_bg, this theme uses dark axis colors to
improve black and white printing or copying.
}
